kafka-consumer/requirements.txt
kafka-python
prometheus_client
python-dotenv
